+++ public/controllers/agent/components/register-agent.js
@@ -110,6 +110,10 @@ const sysButtons = [
   }
 ];
 
+const pTextCheckConnectionStyle = {
+  marginTop: '3em',
+};
+
 export class RegisterAgent extends Component {
   constructor(props) {
     super(props);
@@ -123,6 +127,7 @@ export class RegisterAgent extends Component {
       selectedArchitecture: '',
       selectedVersion: '',
       version: '',
+      wazuhVersion: '',
       serverAddress: '',
       wazuhPassword: '',
       groups: [],
@@ -357,6 +362,9 @@ export class RegisterAgent extends Component {
   };
 
   render() {
+
+    const urlCheckConnectionDocumentation = `https://documentation.wazuh.com/${this.state.wazuhVersion.slice(0,3)}/user-manual/agents/agent-connection.html`;
+    const textAndLinkToCheckConnectionDocumentation = <p style={pTextCheckConnectionStyle}>To verify the connection with the Manager, please follow this <a href={urlCheckConnectionDocumentation} target="_blank">document.</a></p>
     const missingOSSelection = this.checkMissingOSSelection();
     const ipInput = (
       <EuiText>
@@ -423,7 +431,6 @@ export class RegisterAgent extends Component {
       </>
     );
     const restartAgentCommand = this.restartAgentCommand[this.state.selectedOS];
-
     const onTabClick = (selectedTab) => {
       this.selectSYS(selectedTab.id);
     };
@@ -473,6 +480,7 @@ export class RegisterAgent extends Component {
                   </EuiButton>
                 )}
               </EuiCopy>
+              {textAndLinkToCheckConnectionDocumentation}
             </EuiText>
           </Fragment>
         ),
@@ -497,6 +505,7 @@ export class RegisterAgent extends Component {
                   </EuiButton>
                 )}
               </EuiCopy>
+              {textAndLinkToCheckConnectionDocumentation}
             </EuiText>
           </Fragment>
         ),
@@ -604,7 +613,6 @@ export class RegisterAgent extends Component {
       ] : [])
 
     ];
-
     return (
       <div>
         <EuiPage restrictWidth="1000px" style={{ background: 'transparent' }}>